requirements.txt
fastapi
uvicorn
pyyaml
requests
apscheduler
feedparser

# Optional dependencies:
gspread
newsapi-python
oauth2client
